fix(RRB): Minor fixes in rolling red black (#3665)
* Correctly considers instances during disable in RRB if they are not part of discovery
* Better math (e.g. 20% of 5 = 1 not 2!)

+++ clouddriver-core/src/main/groovy/com/netflix/spinnaker/clouddriver/helpers/EnableDisablePercentageCategorizer.groovy
@@ -35,7 +35,7 @@ class EnableDisablePercentageCategorizer<T> {
     }
 
     int totalSize = modified.size() + unmodified.size()
-    int newSize = (int) Math.ceil(totalSize * (float) (desiredPercentage / 100))
+    int newSize = (int) Math.ceil(totalSize * desiredPercentage / 100.0)
 
     int returnSize = modified.size() > newSize ? 0 : newSize - modified.size()
